requirements.txt
fastapi>=0.116.0
uvicorn>=0.34.0
streamlit>=1.49.0
yfinance>=0.2.65
requests>=2.32.0
python-dotenv>=1.1.0
openai>=1.84.0
pandas>=2.3.0
numpy>=1.26.0
beautifulsoup4>=4.13.0
aiohttp>=3.11.0
matplotlib>=3.9.0
plotly>=6.3.0
joblib>=1.4.0
scikit-learn>=1.5.0
python-multipart>=0.0.6
jinja2>=3.1.2
pyyaml>=6.0.0
pathlib>=1.0.0
